app/requirements.txt
uvicorn[standard]==0.15.0
gunicorn==20.1.0
fastapi==0.68.1
# uvloop==0.15.2
# fastapi-utils
google-cloud-aiplatform
merlin-models
nvtabular
gcsfs
google-cloud-storage
